Refactored more classes

+++ DS4Windows/DS4Forms/ViewModels/UpdaterWindowViewModel.cs
@@ -2,25 +2,30 @@
 using System.Collections.Generic;
 using System.Globalization;
 using System.IO;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows.Documents;
+using System.Net.Http;
 using System.Text.Json;
 using System.Text.Json.Serialization;
+using System.Windows.Documents;
+using DS4Windows;
 using HttpProgress;
 using MarkdownEngine = MdXaml.Markdown;
 
 namespace DS4WinWPF.DS4Forms.ViewModels
 {
-    class UpdaterWindowViewModel
+    internal class UpdaterWindowViewModel
     {
-        public const string CHANGELOG_URI = "https://raw.githubusercontent.com/Ryochan7/DS4Windows/jay/DS4Windows/Changelog.min.json";
-
-        private string newversion;
-        public string Newversion { get => newversion; }
-
         private FlowDocument changelogDocument;
+
+
+        public UpdaterWindowViewModel(string newversion)
+        {
+            BuildTempDocument("Retrieving changelog info.Please wait...");
+            Newversion = newversion;
+            //RetrieveChangelogInfo();
+        }
+
+        public string Newversion { get; }
+
         public FlowDocument ChangelogDocument
         {
             get => changelogDocument;
@@ -31,83 +36,78 @@ namespace DS4WinWPF.DS4Forms.ViewModels
                 ChangelogDocumentChanged?.Invoke(this, EventArgs.Empty);
             }
         }
+
         public event EventHandler ChangelogDocumentChanged;
-
-
-        public UpdaterWindowViewModel(string newversion)
-        {
-            BuildTempDocument("Retrieving changelog info.Please wait...");
-            this.newversion = newversion;
-            //RetrieveChangelogInfo();
-        }
 
         public async void RetrieveChangelogInfo()
         {
             // Sorry other devs, gonna have to find your own server
-            Uri url = new Uri(CHANGELOG_URI);
-            string filename = Path.Combine(Path.GetTempPath(), "Changelog.min.json");
-            bool readFile = false;
+            var url = new Uri(Constants.ChangelogUri);
+            var filename = Path.Combine(Path.GetTempPath(), "Changelog.min.json");
+            var readFile = false;
             using (var downloadStream = new FileStream(filename, FileMode.Create))
             {
-                Task<System.Net.Http.HttpResponseMessage> temp = App.requestClient.GetAsync(url.ToString(), downloadStream);
+                var temp = App.requestClient.GetAsync(url.ToString(), downloadStream);
                 try
                 {
                     await temp.ConfigureAwait(true);
                     if (temp.Result.IsSuccessStatusCode) readFile = true;
                 }
-                catch (System.Net.Http.HttpRequestException) { }
+                catch (HttpRequestException)
+                {
+                }
             }
 
-            bool fileExists = File.Exists(filename);
+            var fileExists = File.Exists(filename);
             if (fileExists && readFile)
             {
-                string temp = File.ReadAllText(filename).Trim();
+                var temp = File.ReadAllText(filename).Trim();
                 try
                 {
-                    JsonSerializerOptions options = new JsonSerializerOptions()
+                    var options = new JsonSerializerOptions
                     {
-                        PropertyNameCaseInsensitive = true,
+                        PropertyNameCaseInsensitive = true
                     };
                     options.Converters.Add(new DateTimeJsonConverter.DateTimeConverterUsingDateTimeParse());
-                    ChangelogInfo tempInfo = JsonSerializer.Deserialize<ChangelogInfo>(temp, options);
+                    var tempInfo = JsonSerializer.Deserialize<ChangelogInfo>(temp, options);
                     BuildChangelogDocument(tempInfo);
                 }
-                catch (JsonException) { }
+                catch (JsonException)
+                {
+                }
             }
             else if (!readFile)
             {
                 BuildTempDocument("Failed to retrieve information");
             }
 
-            if (fileExists)
-            {
-                File.Delete(filename);
-            }
+            if (fileExists) File.Delete(filename);
         }
 
         private void BuildChangelogDocument(ChangelogInfo tempInfo)
         {
-            MarkdownEngine engine = new MarkdownEngine();
-            FlowDocument flow = new FlowDocument();
-            foreach (ChangeVersionInfo versionInfo in tempInfo.Changelog.Versions)
+            var engine = new MarkdownEngine();
+            var flow = new FlowDocument();
+            foreach (var versionInfo in tempInfo.Changelog.Versions)
             {
-                ulong versionNumber = versionInfo.VersionNumberInfo.GetVersionNumber();
-                if (versionNumber > DS4Windows.Global.ExecutableVersionLong)
+                var versionNumber = versionInfo.VersionNumberInfo.GetVersionNumber();
+                if (versionNumber > Global.ExecutableVersionLong)
                 {
-                    VersionLogLocale tmpLog = versionInfo.ApplicableInfo(DS4Windows.Global.Instance.Config.UseLang);
+                    var tmpLog = versionInfo.ApplicableInfo(Global.Instance.Config.UseLang);
                     if (tmpLog != null)
                     {
-                        Paragraph tmpPar = new Paragraph();
-                        string tmp = tmpLog.Header;
+                        var tmpPar = new Paragraph();
+                        var tmp = tmpLog.Header;
                         tmpPar.Inlines.Add(new Run(tmp) { Tag = "Header" });
                         flow.Blocks.Add(tmpPar);
 
                         tmpPar.Inlines.Add(new LineBreak());
-                        tmpPar.Inlines.Add(new Run(versionInfo.ReleaseDate.ToUniversalTime().ToString("r")) { Tag = "ReleaseDate" });
+                        tmpPar.Inlines.Add(new Run(versionInfo.ReleaseDate.ToUniversalTime().ToString("r"))
+                            { Tag = "ReleaseDate" });
 
                         tmpLog.BuildDisplayText();
 
-                        FlowDocument tmpDoc = engine.Transform(tmpLog.DisplayLogText);
+                        var tmpDoc = engine.Transform(tmpLog.DisplayLogText);
                         flow.Blocks.AddRange(new List<Block>(tmpDoc.Blocks));
 
                         tmpPar = new Paragraph();
@@ -121,152 +121,113 @@ namespace DS4WinWPF.DS4Forms.ViewModels
 
         private void BuildTempDocument(string message)
         {
-            FlowDocument flow = new FlowDocument();
+            var flow = new FlowDocument();
             flow.Blocks.Add(new Paragraph(new Run(message)));
             ChangelogDocument = flow;
         }
 
         public void SetSkippedVersion()
         {
-            if (!string.IsNullOrEmpty(newversion))
-            {
-                DS4Windows.Global.Instance.LastVersionChecked = newversion;
-            }
+            if (!string.IsNullOrEmpty(Newversion)) Global.Instance.LastVersionChecked = Newversion;
         }
 
         public void BlankSkippedVersion()
         {
-            DS4Windows.Global.Instance.LastVersionChecked = string.Empty;
+            Global.Instance.LastVersionChecked = string.Empty;
         }
     }
 
     public class ChangelogInfo
     {
-        private string latestVersion;
-        private ChangeVersionNumberInfo latestVersionInfo;
-        private DateTime updatedAt;
-        private ChangelogVersions changelog;
+        [JsonPropertyName("latest_version")] 
+        public string LatestVersion { get; set; }
+        
+        [JsonPropertyName("updated_at")] 
+        public DateTime UpdatedAt { get; set; }
 
-        [JsonPropertyName("latest_version")]
-        public string LatestVersion { get => latestVersion; set => latestVersion = value; }
-
-
-        [JsonPropertyName("updated_at")]
-        public DateTime UpdatedAt { get => updatedAt; set => updatedAt = value; }
-
-        [JsonPropertyName("changelog")]
-        public ChangelogVersions Changelog { get => changelog; set => changelog = value; }
+        [JsonPropertyName("changelog")] 
+        public ChangelogVersions Changelog { get; set; }
 
         [JsonPropertyName("latest_version_number_info")]
         [JsonIgnore(Condition = JsonIgnoreCondition.Never)]
-        public ChangeVersionNumberInfo LatestVersionInfo
-        {
-            get => latestVersionInfo;
-            set => latestVersionInfo = value;
-        }
+        public ChangeVersionNumberInfo LatestVersionInfo { get; set; }
     }
 
     public class ChangeVersionNumberInfo
     {
-        private ushort majorPart;
-        private ushort minorPart;
-        private ushort buildPart;
-        private ushort privatePart;
+        [JsonPropertyName("majorPart")] 
+        public ushort MajorPart { get; set; }
 
-        [JsonPropertyName("majorPart")]
-        public ushort MajorPart { get => majorPart; set => majorPart = value; }
+        [JsonPropertyName("minorPart")] 
+        public ushort MinorPart { get; set; }
 
-        [JsonPropertyName("minorPart")]
-        public ushort MinorPart { get => minorPart; set => minorPart = value; }
+        [JsonPropertyName("buildPart")] 
+        public ushort BuildPart { get; set; }
 
-        [JsonPropertyName("buildPart")]
-        public ushort BuildPart { get => buildPart; set => buildPart = value; }
-
-        [JsonPropertyName("privatePart")]
-        public ushort PrivatePart { get => privatePart; set => privatePart = value; }
+        [JsonPropertyName("privatePart")] 
+        public ushort PrivatePart { get; set; }
 
         public ulong GetVersionNumber()
         {
-            ulong temp = (ulong)majorPart << 48 | (ulong)minorPart << 32 |
-                (ulong)buildPart << 16 | privatePart;
+            var temp = ((ulong)MajorPart << 48) | ((ulong)MinorPart << 32) |
+                       ((ulong)BuildPart << 16) | PrivatePart;
             return temp;
         }
     }
 
     public class ChangelogVersions
     {
-        private List<ChangeVersionInfo> versions;
-
-        [JsonPropertyName("versions")]
-        public List<ChangeVersionInfo> Versions { get => versions; set => versions = value; }
+        [JsonPropertyName("versions")] 
+        public List<ChangeVersionInfo> Versions { get; set; }
     }
 
     public class ChangeVersionInfo
     {
-        private string version;
-        private ChangeVersionNumberInfo versionNumberInfo;
-        private string baseHeader;
-        private DateTime releaseDate;
-        private List<VersionLogLocale> versionLocales;
+        [JsonPropertyName("version_str")] 
+        public string Version { get; set; }
 
-        [JsonPropertyName("version_str")]
-        public string Version { get => version; set => version = value; }
+        [JsonPropertyName("base_header")] 
+        public string BaseHeader { get; set; }
 
-        [JsonPropertyName("base_header")]
-        public string BaseHeader { get => baseHeader; set => baseHeader = value; }
+        [JsonPropertyName("release_date")] 
+        public DateTime ReleaseDate { get; set; }
 
-        [JsonPropertyName("release_date")]
-        public DateTime ReleaseDate { get => releaseDate; set => releaseDate = value; }
-
-        [JsonPropertyName("locales")]
-        public List<VersionLogLocale> VersionLocales { get => versionLocales; set => versionLocales = value; }
+        [JsonPropertyName("locales")] 
+        public List<VersionLogLocale> VersionLocales { get; set; }
 
         [JsonPropertyName("version_number_info")]
         [JsonIgnore(Condition = JsonIgnoreCondition.Never)]
-        public ChangeVersionNumberInfo VersionNumberInfo
-        {
-            get => versionNumberInfo; set => versionNumberInfo = value;
-        }
+        public ChangeVersionNumberInfo VersionNumberInfo { get; set; }
 
         public VersionLogLocale ApplicableInfo(string culture)
         {
-            Dictionary<string, VersionLogLocale> tempDict =
+            var tempDict =
                 new Dictionary<string, VersionLogLocale>();
 
-            foreach (VersionLogLocale logLoc in versionLocales)
-            {
-                tempDict.Add(logLoc.Code, logLoc);
-            }
+            foreach (var logLoc in VersionLocales) tempDict.Add(logLoc.Code, logLoc);
 
             VersionLogLocale result = null;
             CultureInfo hairyLegs = null;
             try
             {
-                if (!string.IsNullOrEmpty(culture))
-                {
-                    hairyLegs = CultureInfo.GetCultureInfo(culture);
-                }
+                if (!string.IsNullOrEmpty(culture)) hairyLegs = CultureInfo.GetCultureInfo(culture);
             }
-            catch (CultureNotFoundException) { }
+            catch (CultureNotFoundException)
+            {
+            }
 
             if (hairyLegs != null)
             {
                 if (tempDict.ContainsKey(hairyLegs.Name))
-                {
                     result = tempDict[hairyLegs.Name];
-                }
                 else if (tempDict.ContainsKey(hairyLegs.TwoLetterISOLanguageName))
-                {
                     result =
                         tempDict[hairyLegs.TwoLetterISOLanguageName];
-                }
             }
 
-            if (result == null && versionLocales.Count > 0)
-            {
+            if (result == null && VersionLocales.Count > 0)
                 // Default to first entry if specific culture info not found
-                result = versionLocales[0];
-            }
+                result = VersionLocales[0];
 
             return result;
         }
@@ -274,41 +235,27 @@ namespace DS4WinWPF.DS4Forms.ViewModels
 
     public class VersionLogLocale
     {
-        private string code;
-        private string header;
-        private List<string> logText;
-        private string editor;
-        private List<string> editorsNote;
-        private DateTime editedAt;
+        public string DisplayLogText { get; private set; }
 
-        private string displayLogText;
-        public string DisplayLogText { get => displayLogText; }
+        public string Code { get; set; }
 
-        public string Code { get => code; set => code = value; }
-        public string Header { get => header; set => header = value; }
+        public string Header { get; set; }
 
-        [JsonPropertyName("log_text")]
-        public List<string> LogText
-        {
-            get => logText;
-            set
-            {
-                logText = value;
-            }
-        }
+        [JsonPropertyName("log_text")] 
+        public List<string> LogText { get; set; }
 
-        [JsonPropertyName("editor")]
-        public string Editor { get => editor; set => editor = value; }
+        [JsonPropertyName("editor")] 
+        public string Editor { get; set; }
 
-        [JsonPropertyName("editors_note")]
-        public List<string> EditorsNote { get => editorsNote; set => editorsNote = value; }
+        [JsonPropertyName("editors_note")] 
+        public List<string> EditorsNote { get; set; }
 
-        [JsonPropertyName("updated_at")]
-        public DateTime EditedAt { get => editedAt; set => editedAt = value; }
+        [JsonPropertyName("updated_at")] 
+        public DateTime EditedAt { get; set; }
 
         public void BuildDisplayText()
         {
-            displayLogText = string.Join("\n", logText);
+            DisplayLogText = string.Join("\n", LogText);
         }
     }
 }
